Move toolbox code to launchpad directory
