+++ src/launchpad/components/Console.js
@@ -1,55 +1,28 @@
 const React = require("react");
 const { DOM: dom, PropTypes, createFactory } = React;
 
-const { bindActionCreators, combineReducers } = require("redux");
-const ReactDOM = require("react-dom");
+const { MODE } = require("../../reps/constants");
+const Rep = createFactory(require("../../reps/rep"));
+const Grip = require("../../reps/grip");
 
-const { MODE } = require("./reps/constants");
-const Rep = createFactory(require("./reps/rep"));
-const Grip = require("./reps/grip");
-
-
-// require("./webconsole.css")
-require("./reps.css")
-require("./toolbox.css")
-
-const { renderRoot, bootstrap, L10N } = require("devtools-launchpad");
-const { getValue, isFirefoxPanel } = require("devtools-config");
-
-if (!isFirefoxPanel()) {
-  L10N.setBundle(require("./strings.js"));
-  window.l10n = L10N;
-}
-
-function onConnect({client} = {}) {
-  if (!client) {
-    return;
-  }
-
-  ReactDOM.render(
-    (React.createFactory(RepsConsole))({client}),
-    root
-  );
-}
-
-RepsConsole = React.createClass({
-  getInitialState: function() {
+Console = React.createClass({
+  getInitialState: function () {
     return {
       expressions: []
     };
   },
 
   propTypes: {
-    client: React.PropTypes.object.isRequired
+    client: PropTypes.object.isRequired
   },
 
-  onSubmitForm: function(e) {
+  onSubmitForm: function (e) {
     e.preventDefault();
     let data = new FormData(e.target);
     let expression = data.get("expression");
     this.props.client.clientCommands.evaluate(expression, {})
       .then(result => {
-        this.setState(function(prevState, props) {
+        this.setState(function (prevState, props) {
           return {
             expressions: [{
               input: expression,
@@ -63,13 +36,13 @@ RepsConsole = React.createClass({
       });
   },
 
-  renderRepInAllModes: function({object}) {
+  renderRepInAllModes: function ({object}) {
     return Object.keys(MODE).map(modeKey =>
        this.renderRep({ object, modeKey })
      );
   },
 
-  renderRep: function({ object, modeKey }) {
+  renderRep: function ({ object, modeKey }) {
     return dom.div(
       {
         className: `rep-element ${modeKey}`,
@@ -80,11 +53,11 @@ RepsConsole = React.createClass({
     );
   },
 
-  render: function() {
+  render: function () {
     return dom.main({},
       dom.form({
-          onSubmit: this.onSubmitForm,
-        },
+        onSubmit: this.onSubmitForm,
+      },
         dom.input({
           type: "text",
           placeholder: "Enter an expression",
@@ -94,9 +67,9 @@ RepsConsole = React.createClass({
       dom.div({className: "results"},
         this.state.expressions.map(expression =>
           dom.div({
-              className: "rep-row",
-              key: JSON.stringify(expression)
-            },
+            className: "rep-row",
+            key: JSON.stringify(expression)
+          },
             dom.div({className: "rep-input"}, expression.input),
             dom.div({className: "reps"},
               this.renderRepInAllModes({
@@ -110,8 +83,4 @@ RepsConsole = React.createClass({
   }
 });
 
-let root = document.createElement("div")
-root.innerText = "Waiting for connection";
-
-bootstrap(React, ReactDOM, root)
-  .then(onConnect);
+module.exports = Console;
